chore(datagrid): inline single-caller TypePicker and drop TableViewCoordinating protocol
